Extract argon2 code into crypto service
